refactor(worker-util): move hwinfo to venus-worker-util
*  move hwinfo to venus-worker-util
* add CI for venus-worker-util
