+++ Makefile
@@ -26,14 +26,17 @@ build-worker:
 check-worker:
 	$(MAKE) -C ./venus-worker/ check-all
 
+test-worker-util:
+	$(MAKE) -C ./venus-worker-util/ test-all
+
 build-worker-util:
 	mkdir -p ./dist/bin/
-	rm -rf ./dist/bin/hwinfo
-	$(MAKE) -C ./hwinfo/ build-all
-	cp $(shell cargo metadata --format-version=1 --manifest-path=./hwinfo/Cargo.toml | jq -r ".target_directory")/release/hwinfo ./dist/bin/
+	rm -rf ./dist/bin/venus-worker-util
+	$(MAKE) -C ./venus-worker-util/ build-all
+	cp $(shell cargo metadata --format-version=1 --manifest-path=./venus-worker-util/Cargo.toml | jq -r ".target_directory")/release/venus-worker-util ./dist/bin/
 
 check-worker-util:
-	$(MAKE) -C ./hwinfo/ check-all
+	$(MAKE) -C ./venus-worker-util/ check-all
 
 check-git:
 	./scripts/check-git-dirty.sh
@@ -50,4 +53,4 @@ dev-env:
 	ln -s ../../.githooks/pre-push ./.git/hooks/pre-push
 	$(MAKE) -C ./venus-sector-manager/ dev-env
 	$(MAKE) -C ./venus-worker/ dev-env
-	$(MAKE) -C ./hwinfo/ dev-env
+	$(MAKE) -C ./venus-worker-util/ dev-env
