+++ venus-worker-util/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hwinfo"
+name = "venus-worker-util"
 version = "0.1.0"
 edition = "2021"
