+++ venus-worker-util/src/lib.rs
@@ -1,0 +1,5 @@
+#![deny(missing_docs)]
+//! venus-worker utility collection
+
+/// hardware information
+pub mod hwinfo;
